Darja had cool idea. Skip every other pixel when calculating indirect lighting. Did not really help with optimization but looks better!

+++ Source/Game/Source/Game.h
@@ -152,7 +152,7 @@ namespace Kaka
 			BOOL useDirectionalRSM = true;
 			BOOL useSpotRSM = true;
 			BOOL onlyRSM = false;
-			unsigned int uvScale = 1;
+			float uvScale = 1;
 		};
 
 		RSMBuffer rsmBufferDirectional = {};
